moved last updated session component to standAloneComponents directory
