+++ jada-app/src/Components/Pages/TablesPage/TablesPage.js
@@ -1,7 +1,7 @@
 import React from "react";
 import './tablesPage.css';
 import PageHeader from "../../PageHeader/PageHeader";
-import LastUpdatedSession from "../../LastUpdatedSession/LastUpdatedSession";
+import LastUpdatedSession from "../../StandAloneComponents/LastUpdatedSession/LastUpdatedSession";
 import ButtonMain from "../../Buttons/ButtonMain/ButtonMain";
 import TableApplications from "../../Tables/TableApplications/TableApplications";
